+++ src/resources/workers/scripts/deployments.ts
@@ -1,6 +1,7 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
 import { APIResource } from '../../../resource';
+import { isRequestOptions } from '../../../core';
 import * as Core from '../../../core';
 
 export class Deployments extends APIResource {
@@ -34,7 +35,7 @@ export class Deployments extends APIResource {
     params: DeploymentCreateParams,
     options?: Core.RequestOptions,
   ): Core.APIPromise<Deployment> {
-    const { account_id, force, ...body } = params;
+    const { account_id = this._client.accountId, force, ...body } = params;
     return (
       this._client.post(`/accounts/${account_id}/workers/scripts/${scriptName}/deployments`, {
         query: { force },
@@ -59,10 +60,19 @@ export class Deployments extends APIResource {
    */
   list(
     scriptName: string,
-    params: DeploymentListParams,
+    params?: DeploymentListParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<DeploymentListResponse>;
+  list(scriptName: string, options?: Core.RequestOptions): Core.APIPromise<DeploymentListResponse>;
+  list(
+    scriptName: string,
+    params: DeploymentListParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<DeploymentListResponse> {
-    const { account_id } = params;
+    if (isRequestOptions(params)) {
+      return this.list(scriptName, {}, params);
+    }
+    const { account_id = this._client.accountId } = params;
     return (
       this._client.get(
         `/accounts/${account_id}/workers/scripts/${scriptName}/deployments`,
@@ -88,10 +98,24 @@ export class Deployments extends APIResource {
   delete(
     scriptName: string,
     deploymentId: string,
-    params: DeploymentDeleteParams,
+    params?: DeploymentDeleteParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<DeploymentDeleteResponse>;
+  delete(
+    scriptName: string,
+    deploymentId: string,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<DeploymentDeleteResponse>;
+  delete(
+    scriptName: string,
+    deploymentId: string,
+    params: DeploymentDeleteParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<DeploymentDeleteResponse> {
-    const { account_id } = params;
+    if (isRequestOptions(params)) {
+      return this.delete(scriptName, deploymentId, {}, params);
+    }
+    const { account_id = this._client.accountId } = params;
     return this._client.delete(
       `/accounts/${account_id}/workers/scripts/${scriptName}/deployments/${deploymentId}`,
       options,
@@ -114,10 +138,20 @@ export class Deployments extends APIResource {
   get(
     scriptName: string,
     deploymentId: string,
-    params: DeploymentGetParams,
+    params?: DeploymentGetParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<Deployment>;
+  get(scriptName: string, deploymentId: string, options?: Core.RequestOptions): Core.APIPromise<Deployment>;
+  get(
+    scriptName: string,
+    deploymentId: string,
+    params: DeploymentGetParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<Deployment> {
-    const { account_id } = params;
+    if (isRequestOptions(params)) {
+      return this.get(scriptName, deploymentId, {}, params);
+    }
+    const { account_id = this._client.accountId } = params;
     return (
       this._client.get(
         `/accounts/${account_id}/workers/scripts/${scriptName}/deployments/${deploymentId}`,
@@ -216,7 +250,7 @@ export interface DeploymentCreateParams {
   /**
    * Path param: Identifier.
    */
-  account_id: string;
+  account_id?: string;
 
   /**
    * Body param
@@ -260,21 +294,21 @@ export interface DeploymentListParams {
   /**
    * Identifier.
    */
-  account_id: string;
+  account_id?: string;
 }
 
 export interface DeploymentDeleteParams {
   /**
    * Identifier.
    */
-  account_id: string;
+  account_id?: string;
 }
 
 export interface DeploymentGetParams {
   /**
    * Identifier.
    */
-  account_id: string;
+  account_id?: string;
 }
 
 export declare namespace Deployments {
